python-auth/requirements.txt
face_recognition==1.3.0
opencv-python==4.5.5.64
numpy==1.22.3
psycopg2-binary==2.9.3
dlib==19.24.0
imutils==0.5.4
